v2.1.0 beta1 (#341)
* bump versions

* fix log(0) bug

* refactor examples

* add paper banner to home tab (closes #326)

* fix axis in global sequencing homepage cases panel

* add Delta variant to default selected list
