+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "covidcg",
-  "version": "2.0.9",
+  "version": "2.1.0-beta1",
   "description": "",
   "engines": {
     "node": ">=8",
