+++ src/components/Example/examples.js
@@ -1,10 +1,3 @@
-import React from 'react';
-import { observer } from 'mobx-react';
-import styled from 'styled-components';
-import { useStores } from '../../stores/connect';
-import _ from 'underscore';
-import useDimensions from 'react-use-dimensions';
-
 import { getGene } from '../../utils/gene_protein';
 import { getLocationByNameAndLevel } from '../../utils/location';
 import { queryPrimers } from '../../utils/primer';
@@ -20,14 +13,6 @@ import {
 } from '../../constants/defs.json';
 import { config } from '../../config';
 
-import ExternalLink from '../Common/ExternalLink';
-// import LoadingSpinner from '../Common/LoadingSpinner';
-
-import SurveillancePlot from '../Vega/SurveillancePlot';
-import GlobalSeqPlot from '../Vega/GlobalSeqPlot';
-
-import TempImage from '../../assets/images/cg_short_v13@4x_square.png';
-
 import GlobalLineagesImage from '../../assets/analysis_screens/global_lineages.png';
 // import XinfadiImage from '../../assets/analysis_screens/xinfadi.png';
 import IcelandImage from '../../assets/analysis_screens/iceland.png';
@@ -37,126 +22,8 @@ import USCDCPrimerImage from '../../assets/analysis_screens/us_cdc_primer.png';
 import D614GEuropeNAImage from '../../assets/analysis_screens/d614g_europe_na.png';
 import N203204Image from '../../assets/analysis_screens/n_203_204_coocurrence.png';
 
-const ExampleTabContainer = styled.div`
-  background-color: #f8f8f8;
-`;
-
-const ExampleTabContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-
-  flex-grow: 1;
-  padding: 20px;
-  margin: 0 auto;
-
-  background-color: #fff;
-`;
-
-const ExampleHeader = styled.div`
-  padding-left: 10px;
-  max-width: 800px;
-
-  p {
-    font-weight: normal;
-    line-height: normal;
-  }
-`;
-
-const ExampleTitle = styled.h2``;
-
-const ExampleList = styled.div`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-`;
-const ExampleItem = styled.a`
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-
-  width: 300px;
-  height: 250px;
-  margin: 10px;
-  border: 1px solid #ccc;
-  box-shadow: 0px 3px 4px #aaa;
-
-  text-decoration: none;
-
-  transition: 0.1s all ease-in-out;
-  &:hover {
-    border: 1px solid #00e;
-  }
-`;
-const ExampleItemImage = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-
-  img {
-    width: auto;
-    height: 150px;
-  }
-`;
-const ExampleItemFooter = styled.div`
-  height: 100px;
-  padding: 10px;
-
-  border-top: 1px solid #aaa;
-
-  .example-item-title {
-    display: block;
-    font-size: 1.1em;
-    font-weight: 700;
-    color: #444;
-  }
-
-  .example-item-description {
-    display: block;
-    font-size: 0.85em;
-    font-weight: normal;
-    color: #888;
-    margin: 3px 0px;
-    line-height: normal;
-  }
-  //
-`;
-
-// const TOC = styled.div`
-//   font-weight: normal;
-// `;
-
-// const ExampleTutorial = styled.div`
-//   padding: 10px;
-//   padding-top: 0px;
-//   font-weight: normal;
-//   line-height: normal;
-//   max-width: 800px;
-// `;
-
-// const TutorialImage = styled.img`
-//   width: 750px;
-//   border: 1px solid #aaa;
-// `;
-
-// const scrollToRef = (id, e) => {
-//   if (e !== undefined) {
-//     e.preventDefault();
-//   }
-//   const el = window.document.getElementById(id);
-//   window.scrollTo(0, el.offsetTop);
-// };
-
-const ExampleTab = observer(() => {
-  const { configStore, plotSettingsStore, UIStore, locationDataStore } =
-    useStores();
-
-  const [ref, { width }] = useDimensions();
-
-  const selectTree = locationDataStore.selectTree;
-  const exampleItems = [
+export const getExampleItems = ({ selectTree }) => {
+  return [
     {
       title: 'Global Lineages',
       description: 'View the growth of the B lineage family over all locations',
@@ -390,105 +257,4 @@ const ExampleTab = observer(() => {
       },
     },
   ];
-
-  const onExampleClick = (title, e) => {
-    e.preventDefault();
-
-    const exampleItem = _.findWhere(exampleItems, { title });
-    // console.log(exampleItem);
-
-    // Apply settings for each store
-    Object.keys(exampleItem.settings).forEach((store) => {
-      // Call the example action for each store
-      if (store === 'config') {
-        configStore.resetValues(exampleItem.settings[store]);
-      } else if (store === 'plotSettings') {
-        plotSettingsStore.resetValues(exampleItem.settings[store]);
-      } else if (store === 'UI') {
-        UIStore.resetValues(exampleItem.settings[store]);
-      }
-    });
-  };
-
-  const exampleElements = [];
-  exampleItems.forEach((exampleItem) => {
-    exampleElements.push(
-      <ExampleItem
-        key={`example-item-${exampleItem.title}`}
-        href="#"
-        onClick={onExampleClick.bind(this, exampleItem.title)}
-      >
-        <ExampleItemImage>
-          <img
-            src={
-              exampleItem.image === undefined ? TempImage : exampleItem.image
-            }
-          />
-        </ExampleItemImage>
-        <ExampleItemFooter>
-          <span className="example-item-title">{exampleItem.title}</span>
-          <p className="example-item-description">{exampleItem.description}</p>
-        </ExampleItemFooter>
-      </ExampleItem>
-    );
-  });
-
-  const renderExamples = () => {
-    return exampleElements;
-  };
-
-  return (
-    <ExampleTabContainer ref={ref}>
-      <ExampleTabContent>
-        {/* <TOC>
-          <h3>Table of Contents</h3>
-          <ul>
-            <li>
-              <a
-                href="#introduction"
-                onClick={scrollToRef.bind(this, 'introduction')}
-              >
-                <b>Introduction</b>
-              </a>
-            </li>
-            <li>
-              <a
-                href="#example-analyses"
-                onClick={scrollToRef.bind(this, 'example-analyses')}
-              >
-                <b>Example Analyses</b>
-              </a>
-            </li>
-          </ul>
-        </TOC> */}
-
-        <SurveillancePlot width={width - 150} />
-        <div style={{ height: '15px' }} />
-        <GlobalSeqPlot width={width - 120} />
-
-        <ExampleHeader>
-          {/* <a id="example-analyses" /> */}
-          <ExampleTitle>Example Analyses</ExampleTitle>
-          {/* <a href="#" onClick={scrollToRef.bind(this, 'getting-started-top')}>
-            Back
-          </a> */}
-          <p>
-            Use these example analyses to get started and explore the features
-            of this application. If you would like to add an analysis to this
-            list, please{' '}
-            <ExternalLink href="https://github.com/vector-engineering/covidcg">
-              submit a pull request on our GitHub
-            </ExternalLink>
-            , or contact us at{' '}
-            <ExternalLink href="mailto:covidcg@broadinstitute.org">
-              covidcg@broadinstitute.org
-            </ExternalLink>
-          </p>
-        </ExampleHeader>
-        <ExampleList>{renderExamples()}</ExampleList>
-      </ExampleTabContent>
-    </ExampleTabContainer>
-  );
-});
-
-export default ExampleTab;
+};
