+++ src/components/Pages/HomePage.js
@@ -13,7 +13,7 @@ import KeyListener from '../KeyListener';
 import AsyncErrorModal from '../Modals/AsyncErrorModal';
 
 const CompareGroupsTab = React.lazy(() => import('./CompareGroupsTab'));
-const ExampleTab = React.lazy(() => import('./ExampleTab'));
+const HomeTab = React.lazy(() => import('./HomeTab'));
 const CompareLocationsTab = React.lazy(() => import('./CompareLocationsTab'));
 const GroupReportTab = React.lazy(() => import('./GroupReportTab'));
 const AboutTab = React.lazy(() => import('./AboutTab'));
@@ -48,7 +48,7 @@ const HomePage = observer(() => {
     } else if (UIStore.activeTab === TABS.TAB_GROUP_REPORT) {
       return <GroupReportTab />;
     } else if (UIStore.activeTab === TABS.TAB_EXAMPLE) {
-      return <ExampleTab />;
+      return <HomeTab />;
     } else if (UIStore.activeTab === TABS.TAB_ABOUT) {
       return <AboutTab />;
     } else if (UIStore.activeTab === TABS.TAB_METHODOLOGY) {
