+++ src/stores/groupDataStore.js
@@ -12,7 +12,7 @@ import { mutationHeatmapToPymolScript } from '../utils/pymol';
 
 export const initialValues = {
   activeGroupType: Object.keys(config['group_cols'])[0],
-  selectedGroups: ['B.1.1.7', 'B.1.351', 'P.2'],
+  selectedGroups: ['B.1.617.2', 'B.1.1.7', 'B.1.351', 'P.2'],
   groupSnvType: 'protein_aa',
 };
