+++ src/utils/version.js
@@ -1,1 +1,1 @@
-export const version = '2.0.9';
+export const version = '2.1.0-beta1';
